api-queries: consolidate read-list into read-lists
The api-queries split between read-list.ts (list items) and read-lists.ts
(lists themselves) was inconsistent with api-core, which keeps both under
read-lists/. Per api-queries README, the filename should match the
api-core directory name.

Move readListItemsQuery and readListItemsInfiniteQuery into read-lists.ts
and delete the separate read-list.ts file.
